Add directory changes-in-dat-files/apt.dat with two patches for apt.dat
000-general-modify-header.patch:
  indicate that apt.dat has been modified by the FlightGear team

001-CCC5-fix-missing-Unicom-frequency.patch:
  add missing Unicom frequency for CCC5 (Havelock), thanks to wkitty42

The patch files have LF line endings (otherwise this is a mess: some
lines ending with LF and others with CRLF inside the same file). In
order to obtain this, apt.dat has been recoded to LF line endings (Unix
and MacOS X), then the changes have been made, the diff commands run,
and finally the resulting file has been recoded with CRLF line endings
(Windows) before being gzipped and written to FGData. This was done so
in order to preserve compatibility with software that could not handle a
change in $FG_ROOT/Airports/apt.dat.gz's line ending style.