Add script git-date.py in directory 'dev-utils'
This script allows one to easily find commits in several Git
repositories around a given date (in any date format accepted by Git).
The repositories may be given labels for easier identification and can
be defined via a configuration file and/or the command-line arguments.

Option --checkout tells git-date.py to automatically perform a
'git checkout' in each of the repositories at the commits that were
found. Options --show-commits and --show-commits-option instruct it to
run 'git show' for each commit found with the options of your choice
(all options can also be set via a configuration file).

This script can be useful to find when a given regression happened in
FlightGear---or any project that makes use of several Git repositories.
For more details, run 'git-date.py --help'.